refactor: relocate task_events and audit to pitboss-core
`pitboss-web::api::audit` was reaching into `pitboss-cli::audit` and
`pitboss-cli::dispatch::events` to read the run-wide `audit.jsonl`
artifact — two read-only types whose only legitimate consumer outside
the dispatcher is the web console. Pulling them through `pitboss-cli`
forced the web crate to depend on the full dispatcher (its MCP server,
control bridge, process spawner, …) just to deserialize a JSONL row.

Move both modules into `pitboss-core` where the other consumer-shared
reader-side types already live (`pitboss_core::runs`,
`pitboss_core::store`, `pitboss_core::stream`). The cli-side
`crates/pitboss-cli/src/audit.rs` and
`crates/pitboss-cli/src/dispatch/events.rs` shrink to re-export shims,
matching the pattern `pitboss_cli::runs` already follows since #484. All
existing `crate::dispatch::events::*` / `crate::audit::*` paths inside
`pitboss-cli` continue to resolve through the shim — no internal
caller had to change.

`pitboss-web::api::audit` now imports from `pitboss_core::audit` and
`pitboss_core::task_events` directly, dropping two of its
`pitboss_cli::*` reach-ins. `pitboss-cli` still owns the dispatcher
runtime, manifest pipeline, MCP server, control protocol, and live
stream — those remain too entangled with cli-only types
(`ApprovalRule`, the typed `ControlEvent` payloads, etc.) to relocate
in this PR — but the pure-disk audit types had no such ties.

Notes:
- `pitboss-core` has `#![warn(clippy::pedantic)]` while `pitboss-cli`
  does not. The moved modules carry minimal `#![allow(...)]` attrs for
  `must_use_candidate` and `needless_pass_by_value` so the move is
  mechanical — no behavior change. Two trailing-comma `format!` nits
  got fixed in passing.
- The `pitboss audit` subcommand entry (`run`) was moved alongside the
  data types; the CLI binary's audit subcommand handler still routes
  through `pitboss_cli::audit::run` via the re-export shim.

Closes #482
Closes #484
